+++ CHANGELOG.md
@@ -6,9 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.1.2](https://github.com/in2workspace/in2-issuer-ui/releases/tag/v2.1.2)
 ### Added
-- Get and display subject email in credential details page.
+- Get and display email in credential details page.
 - Get and display organization identifier in management page.
-- Add organization identifier filter in managament page.
 
 ### Changed 
 - Changed "create-as-signer" route for "create-on-behalf".