+++ CHANGELOG.md
@@ -6,14 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.1.2](https://github.com/in2workspace/in2-issuer-ui/releases/tag/v2.1.2)
 ### Added
-- Get and display email in credential details page.
+- Get and display contact email in credential details page.
 - Get and display organization identifier in management page.
 
 ### Changed 
 - Changed "create-as-signer" route for "create-on-behalf".
 
 ### Fixed
-- TranslaAdd missing translations in powers component.
+- Add missing translations.
 
 
 ## [2.1.1](https://github.com/in2workspace/in2-issuer-ui/releases/tag/v2.1.1)